feat: приложение запускается наводим порядок в структуре
