--- src/ducks/auth/index.ts
@@ -1,4 +1,0 @@
-import { authSlice } from "./slice";
-export * from "./thunk";
-
-export const authReducer = authSlice.reducer;
