+++ src/features/auth/index.ts
@@ -1,0 +1,3 @@
+export { authSlice } from "./model/slice.ts";
+export * from "./model/slice.ts";
+export * from "./model/thunks.ts";
