+++ src/features/auth/model/index.ts
@@ -1,0 +1,5 @@
+import { authSlice } from "./slice.ts";
+
+export * from "./thunks.ts";
+
+export const authReducer = authSlice.reducer;
