+++ src/widgets/auth-layout/ui/AuthLayout.module.css
