+++ src/widgets/auth-layout/ui/AuthLayout.tsx
