+++ src/widgets/auth-layout/ui/SignInForm.tsx
@@ -1,11 +1,11 @@
-import { VALIDATE_CHAR_FORM_REGISTRATION } from "@/constants/constants";
-import { signIn } from "@/ducks/auth";
-import { useAppDispatch, useAppSelector } from "@/ducks/hooks";
-import { AuthData } from "@/types/types";
 import { App, Button, Form, Input, notification } from "antd";
 import { useEffect, useState } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import style from "./AuthLayout.module.css";
+import { VALIDATE_CHAR_FORM_REGISTRATION } from "@/constants/constants.tsx";
+import { AuthData } from "@/types/types.tsx";
+import { useAppDispatch, useAppSelector } from "@/ducks/hooks.ts";
+import { signIn } from "@/features/auth/index.ts";
 
 const { MIN_CHAR_LOGIN, MAX_CHAR_LOGIN, MIN_CHAR_PASSWORD, MAX_CHAR_PASSWORD } =
   VALIDATE_CHAR_FORM_REGISTRATION;
