+++ vite.config.ts
@@ -8,6 +8,12 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "src"),
+      "@features": path.resolve(__dirname, "./src/features"),
+      "@entities": path.resolve(__dirname, "./src/entities"),
+      "@shared": path.resolve(__dirname, "./src/shared"),
+      "@widgets": path.resolve(__dirname, "./src/widgets"),
+      "@pages": path.resolve(__dirname, "./src/pages"),
+      "@app": path.resolve(__dirname, "./src/app"),
     },
   },
 });
